refactor: simplify parsing logic
