+++ packages/svg/src/index.ts
@@ -1,3 +1,2 @@
 export { parseSvg } from './svg';
-export type { ParsedSvg } from './svg';
-export type { SvgNode, Viewbox } from './types';
+export type { SvgNode } from './types';
