+++ packages/svg/src/types.ts
@@ -1,10 +1,3 @@
-export type Viewbox = {
-  minX: number;
-  minY: number;
-  maxX: number;
-  maxY: number;
-};
-
 export interface SvgNode {
   type: string;
   props: Record<string, unknown>;
